# Conflicts:
#	log/logback_info.log
#	log/logback_warn.log
#	src/main/java/com/phonecard/dao/AddressMapper.java
#	src/main/java/com/phonecard/dao/AddressSelfMapper.java
#	src/main/java/com/phonecard/dao/CommentMapper.java
#	src/main/java/com/phonecard/dao/FeedbackMapper.java
#	src/main/java/com/phonecard/dao/KefuMapper.java

+++ src/main/java/com/phonecard/form/AdvertisementForm.java
@@ -19,7 +19,7 @@ public class AdvertisementForm {
 
     private Short type;
 
-    private Integer isShow;
+    private Short isShow;
 
     private Short isDelete;
 }
